confluent-agents/requirements.txt
confluent-kafka==2.3.0
python-dotenv==1.0.0
requests==2.31.0
fastmcp==0.2.0